DS4300_Final_Project-main/requirements.txt
streamlit
boto3
python-dotenv
pymysql
seaborn